# Refactor: Performance improvements