… .pre-commit-config.yaml
… ansible_collections/arista/avd/docs/plugins/Filter_plugins/convert_schema.md
… ansible_collections/arista/avd/docs/plugins/Modules_and_action_plugins/batch_template.md
… ansible_collections/arista/avd/docs/plugins/Modules_and_action_plugins/validate_and_template.md
… ansible_collections/arista/avd/playbooks/_build_schemas.yml
… ansible_collections/arista/avd/plugins/filter/convert_schema.py
… ansible_collections/arista/avd/plugins/plugin_utils/schema/avd_meta_schema.pickle
… ansible_collections/arista/avd/plugins/plugin_utils/schema/avd_meta_schema.schema.pickle
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/aaa-accounting.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/aaa-authentication.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/aaa-authorization.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/aaa-root.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/aaa-server-groups.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/access-lists.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/address-locking.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/agents.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/aliases.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/application-traffic-recognition.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/arp.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/as-path.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/banners.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/bgp-groups.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/boot.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/class-maps.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/clock.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/community-lists.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/custom-templates.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/cvx.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/daemon-terminattr.md
… ansible_collections/arista/avd/roles/eos_cli_config_gen/docs/tables/daemons.md