## Change Summary

<!-- Enter short PR description -->
Performance improvements in eos_designs and common schema tooling

## Proposed changes
<!--- Describe your changes in detail -->
<!--- Describe data model implemented for new features -->

Based on cProfiling this focuses on the most time impacting areas

- Avoid natural sorting internal data structured when not needed.
- Move schema build to Python instead of Ansible.
- Compile a fully resolved schema and store as "pickle" when building schemas.
- Load schema store from fully-resolved pickles instead of YAML with $ref.
- Adjust schema tooling to not run through resolving schemas during validation

## How to test
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->

- No changes to molecule outputs (so conversion works as expected)
- Runtime is better (390 devices, 12 forks):
  ```txt
  arista.avd.eos_cli_config_gen : Generate eos intended configuration---------------- 488.45s -> 473.17s
  arista.avd.eos_designs : Generate device configuration in structured format ------- 558.30s -> 149.62s
  arista.avd.eos_designs : Set eos_designs facts ------------------------------------ 199.58s -> 46.85s
  ```

## Checklist

### User Checklist

<!-- Add your own checklist using MD syntax and by replacing N/A -->
- N/A

### Repository Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code has been rebased from devel before I start
- [x] I have read the [**CONTRIBUTING**](https://avd.sh/en/latest/docs/contribution/overview.html) document.
- [x] My change requires a change to the documentation and documentation have been updated accordingly.
- [x] I have updated [molecule CI](https://github.com/aristanetworks/avd/tree/devel/ansible_collections/arista/avd/molecule) testing accordingly. (check the box if not applicable)